removed check-style plug-in and other redundant code

+++ src/main/resources/init_db.sql
@@ -1,6 +1,6 @@
 CREATE TABLE `books` (
  `id` BIGINT NOT NULL AUTO_INCREMENT,
  `title` VARCHAR(255),
- `price` BIGINT,
+ `price` DECIMAL,
  PRIMARY KEY (`id`)
  );
